Add asserts to make sure columns used by property filters are indexed. Also move metadata + asserts logic to separate files
