+++ ogc/features/datasources/geopackage/asserts.go
@@ -1,0 +1,69 @@
+package geopackage
+
+import (
+	"fmt"
+	"strings"
+
+	"github.com/PDOK/gokoala/engine"
+	"github.com/jmoiron/sqlx"
+)
+
+// assert required indexes in the GeoPackage exists
+func assertIndexesExist(
+	configuredCollections engine.GeoSpatialCollections,
+	featureTableByCollectionID map[string]*featureTable,
+	db *sqlx.DB, fidColumn string) error {
+
+	// index needs to contain these columns in the given order
+	spatialBtreeColumns := strings.Join([]string{fidColumn, "minx", "maxx", "miny", "maxy"}, ",")
+
+	for collID, table := range featureTableByCollectionID {
+		for _, coll := range configuredCollections {
+			if coll.ID == collID && coll.Features != nil {
+				// assert spatial b-tree index exists, this index substitutes the r-tree when querying large bounding boxes
+				if err := assertIndexExists(table, db, spatialBtreeColumns); err != nil {
+					return err
+				}
+
+				// assert the column for each property filter is indexed.
+				for _, propertyFilter := range coll.Features.Filters.Properties {
+					if err := assertIndexExists(table, db, propertyFilter.Name); err != nil {
+						return err
+					}
+				}
+				break
+			}
+		}
+	}
+	return nil
+}
+
+func assertIndexExists(table *featureTable, db *sqlx.DB, columns string) error {
+	if table == nil {
+		return fmt.Errorf("given table can't be nil")
+	}
+	query := fmt.Sprintf(`
+select group_concat(info.name) as indexed_columns
+from pragma_index_list('%s') as list,
+     pragma_index_info(list.name) as info
+group by list.name`, table.TableName)
+
+	rows, err := db.Queryx(query)
+	if err != nil {
+		return fmt.Errorf("failed to read indexes from table '%s'", table.TableName)
+	}
+	exists := false
+	for rows.Next() {
+		var indexedColumns string
+		_ = rows.Scan(&indexedColumns)
+		if columns == indexedColumns {
+			exists = true
+		}
+	}
+	rows.Close() //nolint:sqlclosecheck // don't defer in loop
+	if !exists {
+		return fmt.Errorf("missing required index: no index exists on column(s) '%s' in table '%s'",
+			columns, table.TableName)
+	}
+	return nil
+}
